원정대 모달 추가 . 경험치 획득 경로 추가 (미션, 캐릭터뽑기)

+++ src/main/java/com/phobi/gamja/entity/contents/CorpsTier.java
@@ -14,7 +14,7 @@ public class CorpsTier {
 
     @Id
     @Column(name = "tier_id")
-    private Integer tierId;
+    private Long tierId;
 
     @Column(name = "name", nullable = false)
     private String name;
